requirements.txt
streamlit
torch
torchvision
opencv-python-headless
pillow
pandas
requests
PyYAML
tqdm
seaborn
matplotlib